+++ shuup/xtheme/admin_module/__init__.py
@@ -15,11 +15,13 @@ from django_jinja.backend import Jinja2
 from shuup.admin.base import AdminModule, MenuEntry, Notification
 from shuup.admin.menu import CONTENT_MENU_CATEGORY
 from shuup.admin.utils.permissions import get_default_model_permissions
-from shuup.admin.utils.urls import admin_url
+from shuup.admin.utils.urls import (
+    admin_url, derive_model_url, get_edit_and_list_urls
+)
 from shuup.admin.views.home import HelpBlockCategory, SimpleHelpBlock
 from shuup.xtheme._theme import get_current_theme
 from shuup.xtheme.engine import XthemeEnvironment
-from shuup.xtheme.models import ThemeSettings
+from shuup.xtheme.models import Snippet, ThemeSettings
 
 
 class XthemeAdminModule(AdminModule):
@@ -46,7 +48,7 @@ class XthemeAdminModule(AdminModule):
                 permissions=get_default_model_permissions(ThemeSettings)
             ),
             admin_url(
-                "^xtheme/",
+                "^xtheme/theme",
                 "shuup.xtheme.admin_module.views.ThemeConfigView",
                 name="xtheme.config",
                 permissions=get_default_model_permissions(ThemeSettings)
@@ -96,3 +98,45 @@ class XthemeAdminModule(AdminModule):
                         title=_("Theming"),
                         url="shuup_admin:xtheme.config"
                     )
+
+
+class XthemeSnippetsAdminModule(AdminModule):
+    name = _("Shuup Extensible Theme Engine Snippets")
+    breadcrumbs_menu_entry = MenuEntry(
+        _("Snippets"),
+        "shuup_admin:xtheme_snippet.list",
+        category=CONTENT_MENU_CATEGORY
+    )
+
+    def get_urls(self):
+        return get_edit_and_list_urls(
+            url_prefix="^xtheme/snippet",
+            view_template="shuup.xtheme.admin_module.views.Snippet%sView",
+            name_template="xtheme_snippet.%s",
+            permissions=get_default_model_permissions(Snippet)
+        ) + [
+            admin_url(
+                "^xtheme/snippet/(?P<pk>\d+)/delete/$",
+                "shuup.xtheme.admin_module.views.SnippetDeleteView",
+                name="xtheme_snippet.delete",
+                permissions=get_default_model_permissions(Snippet)
+            )
+        ]
+
+    def get_menu_entries(self, request):
+        return [
+            MenuEntry(
+                text=_("Theme Snippet Injection"),
+                icon="fa fa-magic",
+                url="shuup_admin:xtheme_snippet.list",
+                category=CONTENT_MENU_CATEGORY,
+                subcategory="design",
+                ordering=2
+            )
+        ]
+
+    def get_required_permissions(self):
+        return get_default_model_permissions(Snippet)
+
+    def get_model_url(self, object, kind, shop=None):
+        return derive_model_url(Snippet, "shuup_admin:xtheme_snippet", object, kind)